+++ src/layman/common/prime_db_schema/migrate_test.py
@@ -1,7 +1,8 @@
 import test.flask_client as client_util
 
 from layman import settings, app as app
-from . import ensure_schema, migrate_users_and_publications, model, publications
+from . import model, publications as pub_util, workspaces as workspaces_util
+from .schema_initialization import migrate_users_and_publications, ensure_schema
 from .util import run_query, run_statement
 from layman import util
 from layman.layer import util as layer_util
@@ -18,7 +19,9 @@ def test_recreate_schema(client):
 
     with app.app_context():
         run_statement(model.DROP_SCHEMA_SQL)
-        ensure_schema()
+        ensure_schema(settings.LAYMAN_PRIME_SCHEMA,
+                      app,
+                      settings.PUBLICATION_MODULES)
 
     client_util.delete_layer(username, 'test_recreate_schema_layer1', client)
     client_util.delete_map(username, 'test_recreate_schema_map1', client)
@@ -32,17 +35,26 @@ def test_recreate_schema(client):
 
 def test_schema(client):
     username = 'migration_test_user1'
-    client_util.publish_layer(username, 'migration_test_layer1', client)
-    client_util.publish_map(username, 'migration_test_map1', client)
+    layername = 'migration_test_layer1'
+    mapname = 'migration_test_map1'
+    client_util.publish_layer(username, layername, client)
+    client_util.publish_map(username, mapname, client)
 
     with app.app_context():
         run_statement(model.DROP_SCHEMA_SQL)
-        ensure_schema()
-        users = run_query(f'select count(*) from {DB_SCHEMA}.users;')
-        assert users[0][0] == len(util.get_usernames())
+        ensure_schema(settings.LAYMAN_PRIME_SCHEMA,
+                      app,
+                      settings.PUBLICATION_MODULES)
+        workspaces = run_query(f'select count(*) from {DB_SCHEMA}.workspaces;')
+        assert workspaces[0][0] == len(util.get_usernames())
+        user_infos = workspaces_util.get_workspace_infos(username)
+        assert username in user_infos
+        pub_infos = pub_util.get_publication_infos(username)
+        assert layername in pub_infos
+        assert mapname in pub_infos
 
-    client_util.delete_layer(username, 'migration_test_layer1', client)
-    client_util.delete_map(username, 'migration_test_map1', client)
+    client_util.delete_layer(username, layername, client)
+    client_util.delete_map(username, mapname, client)
 
     with app.app_context():
         pubs = layer_util.get_layer_infos(username)
@@ -71,13 +83,13 @@ def test_steps(client):
         exists_right_types = run_query(f'select count(*) from {DB_SCHEMA}.right_types;')
         assert exists_right_types[0][0] == 2
 
-        exists_users = run_query(f'select count(*) from {DB_SCHEMA}.users;')
-        assert exists_users[0][0] == 0
+        exists_workspaces = run_query(f'select count(*) from {DB_SCHEMA}.workspaces;')
+        assert exists_workspaces[0][0] == 0
         exists_pubs = run_query(f'select count(*) from {DB_SCHEMA}.publications;')
         assert exists_pubs[0][0] == 0
-        migrate_users_and_publications()
-        exists_users = run_query(f'select count(*) from {DB_SCHEMA}.users;')
-        assert exists_users[0][0] > 0
+        migrate_users_and_publications(settings.PUBLICATION_MODULES)
+        exists_workspaces = run_query(f'select count(*) from {DB_SCHEMA}.workspaces;')
+        assert exists_workspaces[0][0] > 0
         exists_pubs = run_query(f'select count(*) from {DB_SCHEMA}.publications;')
         assert exists_pubs[0][0] > 0