+++ src/layman/user/util.py
@@ -1,8 +1,8 @@
 from flask import g, current_app
-from layman import LaymanError
+from layman import LaymanError, settings
 from layman.authn import get_open_id_claims, get_iss_id, get_sub
 from layman.util import slugify, to_safe_names, check_username, get_usernames, ensure_whole_user, delete_whole_user
-from layman.authn import redis as authn_redis, filesystem as authn_filesystem
+from layman.authn import redis as authn_redis, filesystem as authn_filesystem, prime_db_schema as authn_prime_db_schema
 
 
 def get_user_profile(user_obj):
@@ -75,6 +75,7 @@ def _save_reservation(username, claims):
     sub = get_sub()
     authn_redis.save_username_reservation(username, iss_id, sub)
     authn_filesystem.save_username_reservation(username, iss_id, sub, claims)
+    authn_prime_db_schema.save_username_reservation(username, iss_id, sub, claims)
     g.user['username'] = username